refactor: index active_conversation_info by email.
The active_converstation_information is a Set now as it used to
store just one key earlier as a dict.
Previously the indexing was done by name which is changed
to email now because of the possibility of non-unique names
present in the active_conversation_info.
In core.py previously there was no need of user_dict as the function
would directly get the full names,but now active_conversation_info
get the full_names of the emails present in self.active_conversation_
info.