views: Add a function _set_inital_focus in TopicsView.
This function sets the inital focus of the topic list.
Instead of writing it directly in the init function where it
is called,it is written as a separate function for easy/clear
testing.
The function loops over self.log but when the desired topic name
is found, self.list_box's position.Instead of using set_focus
function,focus_postion is used for easy testing.