boxes: Create _setup_common_private_compose.
The new approach is similar to how streams setup the coompose box. Here
_setup_common_private_compose setsup the common requirements of both the
case i.e private_box_view and private_box_edit_view(future
commit).Mainly the logic behind the self.to_write_box such as it's
auto-complete and when to send typing event updates(yes in reply state
and no in edit state) is not covered here as they are non-common.