boxes: Disable cycling while editing a Private Message.
The WriteBox has same keypress for the box created for editing as well
as reply. There is no neet to check the validity of self.to_write_box in
case of editing as it's non-editable.So now it checks the msg_edit_state
. Also now if the focus_position is FOCUS_CONTAINER_MESSAGE then the tab
will change the focus to FOCUS_CONTAINER_HEADER only if its a reply
(msg_edit_state =None).
Tests Adapted.